CONN-10463 Reuse MetadataRecord instead of creating a new RecordMetadata class
